# Remove useless Rule::Normal#initialize