@casperisfine 2016-12-23
I forgot, here's the final profile:
```ruby
==================================
  Mode: cpu(1000)
  Samples: 98 (0.00% miss rate)
  GC: 36 (36.73%)
==================================
     TOTAL    (pct)     SAMPLES    (pct)     FRAME
        28  (28.6%)          28  (28.6%)     PublicSuffix::Domain.name_to_labels
        21  (21.4%)          21  (21.4%)     PublicSuffix::Rule::Base#initialize
        34  (34.7%)           6   (6.1%)     block in PublicSuffix::List#reindex!
        26  (26.5%)           5   (5.1%)     #<Module:0x007ffd15212748>.factory
        28  (28.6%)           2   (2.0%)     block (2 levels) in PublicSuffix::List.parse
```

So basically all the time is now spent in allocating objects and `Array#split`, so not likely to be optimizable more.