@weppos 2016-12-26
Hi @casperisfine, thanks for taking the time to profile it!

I definitely agree, given the performance cost I don't see any particular reason to keep the empty initializer. The explicit options may be a maintenance effort, but for now I'm fine to merge the change as I don't plan to alter the API quite often.