@weppos 2016-12-26
I made a change in master that should fix the CI. Do you mind to merge master into this branch just to confirm the CI is green?